[clang][sema] Move the initializer lifetime checking code from SemaInit.cpp to a new place. NFC
This is a refactoring change for better code isolation and reuse, a
first step to extend it for assignments.
